moved ProcosysTagDetailsComparer to sep file
